ISSUE #3168 - fix circular dependency
