+++ frontend/src/v4/helpers/searching.ts
@@ -16,7 +16,7 @@
  */
 
 import { get, groupBy, isArray, map } from 'lodash';
-import { FILTER_TYPES } from '../routes/components/filterPanel/filterPanel.component';
+import { FILTER_TYPES } from '../routes/components/filterPanel/filterPanel';
 
 export const compareStrings = (string1, string2) => {
 	return (string1 || '').toLowerCase().includes((string2 || '').toLowerCase());
